+++ src/sisl/io/vasp/tests/test_chg.py
@@ -62,13 +62,47 @@ def test_nitric_oxide_pol(sisl_files, chg_type, np_dtype):
 
     assert grid.grid.sum() * grid.dvolume == pytest.approx(1, rel=1e-3)
     assert geom == grid.geometry
+    # check against first raw datapoint on grid
+    if chg_type == "CHG":
+        assert grid.grid[0, 0, 0] * grid.volume == pytest.approx(0.33228e-03)
+    elif chg_type == "CHGCAR":
+        assert grid.grid[0, 0, 0] * grid.volume == pytest.approx(0.49880745183e-03)
+
+    # construct up-spin density
+    chg = chgSileVASP(f).read_grid(0, dtype=np_dtype)
+    grid.grid += chg.grid
+    grid.grid /= 2
+    up_spin = chgSileVASP(f).read_grid([0.5, 0.5], dtype=np_dtype)
+    assert np.allclose(grid.grid, up_spin.grid)
 
 
 def test_nitric_oxide_soi(sisl_files, chg_type, np_dtype):
     f = sisl_files(_dir, "nitric_oxide/soi", chg_type + ".gz")
+    grids = []
+    for i in range(4):
+        grid = chgSileVASP(f).read_grid(i, dtype=np_dtype)
+        grids.append(grid)
+        # check against first raw datapoint on grid
+        if chg_type == "CHG":
+            val = [0.12505, 0.26118e-03, 0.26791e-03, 0.27440e-03][i]
+            assert grid.grid[0, 0, 0] * grid.volume == pytest.approx(val)
+        elif chg_type == "CHGCAR":
+            val = [
+                0.12504989341e00,
+                0.26117671499e-03,
+                0.26791165124e-03,
+                0.27440440526e-03,
+            ][i]
+            assert grid.grid[0, 0, 0] * grid.volume == pytest.approx(val)
+
     s = 0
     for i in range(1, 4):
-        grid = chgSileVASP(f).read_grid(i, dtype=np_dtype)
+        grid = grids[i]
         s += (grid.grid.sum() * grid.dvolume) ** 2
-
     assert s == pytest.approx(1, rel=1e-3)
+
+    # construct up-spin density
+    grid.grid += grids[0].grid
+    grid.grid /= 2
+    up_spin = chgSileVASP(f).read_grid([0.5, 0, 0, 0.5], dtype=np_dtype)
+    assert np.allclose(grid.grid, up_spin.grid)